Added time stamp placeholder for exising urls

+++ src/app/core/services/history.service.ts
@@ -7,6 +7,7 @@ import { History } from '../../shared/models/history.model';
 export class HistoryService {
   private qrGenerationHistory = signal<History[]>([]);
   readonly qrHistory = this.qrGenerationHistory.asReadonly();
+  private readonly historyCountLimit = 20;
 
   private readonly localStorageHistoryName = "history";
 
@@ -19,6 +20,27 @@ export class HistoryService {
   }
 
   addHistoryEntry(historyEntry: History) {
+    if (this.qrGenerationHistory().length == this.historyCountLimit) {
+      this.qrGenerationHistory.update(entries => entries.slice(1));
+    }
+
+    let history = localStorage.getItem(this.localStorageHistoryName);
+    let urlExistsInMemory = false;
+    let currentHistoryEntries: History[] = [];
+
+    if (history) {
+      currentHistoryEntries = JSON.parse(history) as History[];
+      urlExistsInMemory = currentHistoryEntries.some(entry => entry.url == historyEntry.url);
+    }
+
+    if (urlExistsInMemory) {
+      currentHistoryEntries = currentHistoryEntries.map(entry => entry.url == historyEntry.url
+        ? { ...entry, creationTime: "Updated time" } : entry);
+
+      this.qrGenerationHistory.set(currentHistoryEntries);
+      return;
+    }
+
     this.qrGenerationHistory.set([...this.qrGenerationHistory(), historyEntry]);
     localStorage.setItem(this.localStorageHistoryName, JSON.stringify(this.qrGenerationHistory()));
   }
